Move showcase page to showcase app (closes #84)
