+++ apps/website/urls.py
@@ -8,8 +8,6 @@ urlpatterns = [
     path("map/<slug:slug>", views.MapPage.as_view(), name="map"),
     path("share", views.SharePage.as_view(), name="share"),
     path("share/<slug:slug>", views.SharePage.as_view(), name="share"),
-    path("showcase", views.ShowcasePage.as_view(), name="showcase"),
-    path("showcase/<slug:slug>", views.ShowcasePage.as_view(), name="showcase"),
     path("privacy", views.PrivacyPage.as_view(), name="privacy"),
     path("robots.txt", views.robots_txt, name="robots"),
 ]
